chore(web): split systemStores.ts
Splits systemStores.ts into stores.ts and platformSystemStore.ts under
engine/js-processor, and systemStores.ts under engine/keyboard, in
preparation for web-core use of system stores.
