+++ web/src/engine/js-processor/src/index.ts
@@ -1,8 +1,8 @@
 export { BeepHandler, JSKeyboardProcessor, LogMessageHandler, ProcessorInitOptions } from "./jsKeyboardProcessor.js";
 export { RuleBehavior } from "./ruleBehavior.js";
 export { JSKeyboardInterface, KeyInformation, StoreNonCharEntry } from "./jsKeyboardInterface.js";
-export * from "./systemStores.js";
 export * from "./deadkeys.js";
+export * from "./stores.js";
 export { OutputTargetBase } from "./outputTargetBase.js";
 export * from "./outputTargetBase.js";
 export { Mock } from "./mock.js";
