+++ web/src/engine/js-processor/src/stores.ts
@@ -1,0 +1,26 @@
+/*
+ * Keyman is copyright (C) SIL Global. MIT License.
+ *
+ * Definitions for basic stores and variable stores (see kmn reference)
+ */
+
+import { StoreNonCharEntry } from './jsKeyboardInterface.js';
+
+/*
+* Type alias definitions to reflect the parameters of the fullContextMatch() callback (KMW 10+).
+* No constructors or methods since keyboards will not utilize the same backing prototype, and
+* property names are shorthanded to promote minification.
+*/
+type PlainKeyboardStore = string;
+
+export type KeyboardStoreElement = (string | StoreNonCharEntry);
+export type ComplexKeyboardStore = KeyboardStoreElement[];
+
+export type KeyboardStore = PlainKeyboardStore | ComplexKeyboardStore;
+
+export type VariableStore = { [name: string]: string };
+
+export interface VariableStoreSerializer {
+  loadStore(keyboardID: string, storeName: string): VariableStore;
+  saveStore(keyboardID: string, storeName: string, storeMap: VariableStore): void;
+}
