+++ web/src/engine/keyboard/src/index.ts
@@ -29,6 +29,7 @@ export { default as KeyEvent } from "./keyEvent.js";
 export * from "./keyEvent.js";
 export { default as KeyMapping } from "./keyMapping.js";
 export { OutputTarget } from "./outputTarget.js";
+export * from "./systemStores.js";
 
 export * from "@keymanapp/web-utils";
